requirements.txt
Django>=3.2
djangorestframework
psycopg2-binary
redis
requests
django-cors-headers
django-environ
celery
python-dotenv

djangorestframework-simplejwt
graphene-django
django-prometheus
